consolidate and document number util functions
